example/nomad_only_for_stateful/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
motor==3.3.2
pymongo==4.5.0
aiokafka==0.8.0
kafka-python==2.0.2
pydantic==1.10.7
python-multipart==0.0.6
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-exporter-jaeger==1.19.0
